moved file and dir methods to file_helper
